% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.datashape.R
\name{multiAPCMC.datashape}
\alias{multiAPCMC.datashape}
\title{multiAPCMC.datashape}
\usage{
multiAPCMC.datashape(
  data,
  noperiod = 4,
  startestage = 5,
  refper1st,
  refper2nd,
  refcoh1st,
  refcoh2nd,
  cases = TRUE
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{noperiod}{the number of periods in the prediction base. For example, if you have 20 years of data,
and data is categorized as 1 year period by 1 year age, then '10' means you only use the last 10 years of the data to extrapolate towards the future.}

\item{startestage}{youngest age group to be included in the regression model. Note that this is based on how data is categorized.
If you have have ages going from 0 to 75 with data on single year ages, then age 0 is the first age category and hence are age '1', age 1 is then '2', and so forth.
If you have 5-year age categories and data starting from age 0 to 4, then '1' is age category 0 to 4, '2' is age category 5 to 9, and so forth.}

\item{refper1st}{first reference category of period, based on 'period codes'; earliest period is 1, second earliest is 2, etc.}

\item{refper2nd}{second reference category of period, based on 'period codes'; earliest period is 1, second earliest is 2, etc.}

\item{refcoh1st}{first reference category of cohort, based on 'cohort codes'; earliest cohort is 1, second earliest is 2, etc.}

\item{refcoh2nd}{second reference category of cohort, based on 'cohort codes'; earliest cohort is 1, second earliest is 2, etc.}

\item{cases}{the name (as a character) of the column containing the cluster identifiers.}
}
\value{
returns a data.frame with a structure that has the right constraints to subsequently be used by multiAPCMC.multifit and multiAPCMC.singlefit
}
\description{
The APCMC approach, following Clayton & Schifflers (Statistics in Medicine, 1987), works by setting additional reference categories on the time dimensions.
This is most easily accomplished by directly manipulating the factor coding for the time dimensions. Datashape does this.
}
\examples{
set.seed(100)
# datashape will generally not be called directly by a user.
# Instead, it will be called by multiAPCMC.multifit.
multiAPCMC.example.data <- multiAPCMC.example.data
apcdata <- multiAPCMC.datashape(multiAPCMC.example.data,
                                noperiod=length(unique(multiAPCMC.example.data$Period)),
                                startestage=1,
                                refper1st=1,
                                refper2nd=2,
                                refcoh1st=1,
                                refcoh2nd=2)

}
