% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.example.data.R
\docType{data}
\name{multiAPCMC.example.data}
\alias{multiAPCMC.example.data}
\title{Example Data for the multiAPCMC package}
\format{
a dataframe with 4 columns, which are:
\describe{
\item{Age}{a variable denoting age, goes from age 0 to age 84.}
\item{Period}{a variable denoting period, goes from the year 1989 to the year 2009.}
\item{PY}{a variable denoting person-years at risk.}
\item{cases}{the number of new 'cases' (e.g. new cancer cases in that year and age).}
}
}
\usage{
multiAPCMC.example.data
}
\description{
A dataframe with artificially generated data, intended to be used in a demonstration of the functions in this package.
}
\keyword{datasets}
