% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.example.futuredata.R
\docType{data}
\name{multiAPCMC.example.futuredata}
\alias{multiAPCMC.example.futuredata}
\title{Example Data for the multiAPCMC package; this is used as 'out of sample' (oos) data}
\format{
a dataframe with 4 columns, which are:
\describe{
\item{Age}{a variable denoting age, goes from age 0 to age 84.}
\item{Period}{a variable denoting period, goes from the year 2010 to the year 2020.}
\item{PY}{a variable denoting person-years at risk.}
\item{cases}{the number of new 'cases' (e.g. new cancer cases in that year and age).}
}
}
\usage{
multiAPCMC.example.futuredata
}
\description{
A dataframe with artificially generated data, intended to be used in a demonstration of the functions in this package. It does not have to represent data that are truly from the future (because then: why would you forecast?) but instead data that are kept back to see which APC model performs best.
}
\keyword{datasets}
