% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.multifit.R
\name{multiAPCMC.multifit}
\alias{multiAPCMC.multifit}
\title{multiAPCMC.multifit}
\usage{
multiAPCMC.multifit(
  data,
  startestage,
  vec.link,
  vec.noperiod,
  vec.refper,
  vec.refcoh,
  nomodelmax = 1024
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{startestage}{youngest age group to be included in the regression model. Note that this is based on how data is categorized. If you have have ages going from 0 to 75 with data on single year ages, then age 0 is the first age category and hence are age '1', age 1 is then '2', and so forth. If you have 5-year age categories and data starting from age 0 to 4, then '1' is age category 0 to 4, '2' is age category 5 to 9, and so forth.}

\item{vec.link}{a vector stating which link functions the user wants to explore; choices are log-link ("log") and power5-link ("power5")}

\item{vec.noperiod}{a vector containing the number of periods in the prediction base that the user wants to explore. For example, if you have 20 years of data, and data is categorized as 1 year period by 1 year age, then '10' means you only use the last 10 years of the data to extrapolate towards the future. Here you can specify multiple options, etc. c(3,5,7,10).}

\item{vec.refper}{which types of references to use for period. Options are 'extremes' (the first and last period), 'outer' (periods that represent roughly the 20\% and 80\% quantiles of all periods), 'center' (the middle categories), 'first middle' (the firsts period and the middle period), 'middle last' (the middle and the last category), 'first second' (the first and second period categories), 'penultimate last' (the penultimate and the last period category).}

\item{vec.refcoh}{which types of references to use for cohort. Options are 'extremes' (the first and last cohort), 'outer' (cohort that represent roughly the 20\% and 80\% quantiles of all cohorts), 'center' (the middle categories), 'first middle' (the firsts cohort and the middle cohort), 'middle last' (the middle and the last category), 'first second' (the first and second cohort categories), 'penultimate last' (the penultimate and the last cohort category).}

\item{nomodelmax}{outputs an error if more models would be fitted than this number.}
}
\value{
returns a nested list with all of the fitted models.
}
\description{
Fits multiple specifications of an APCMC model at once.
}
\examples{
multiAPCMC.example.data <- multiAPCMC.example.data
multiAPCMC.multifit(multiAPCMC.example.data,
                    startestage=5,
                    vec.link=c('log'),
                    vec.noperiod=c(5,10),
                    vec.refper=c('outer','center'),
                    vec.refcoh=c('center'),
                    nomodelmax=1024)

}
