% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.multipred.R
\name{multiAPCMC.multipred}
\alias{multiAPCMC.multipred}
\title{multiAPCMC.multipred}
\usage{
multiAPCMC.multipred(multiAPCMC.multifit.object, futuredata, cuttrend)
}
\arguments{
\item{multiAPCMC.multifit.object}{The list of fitted APCMC models that is produced by \code{multiAPCMC.multifit}.}

\item{futuredata}{a dataframe containing information on future periods and person-years (exposed population) in those years. Should have the same resolution as the data on which the models were also fitted (e.g. 1 year age group by 1 year periods, or 5-year age groups by 5-year periods, and so forth) and should have the columns Period, Age, and PY (person-years).}

\item{cuttrend}{Cut trend in predictions? Should be a vector with at least a length equal to number of periods to predict, e.g. \code{c(0,0,0,0)} or \code{c(0,0.25,0.50,0.75)} or so if we project 4 periods into the future.}
}
\value{
returns a nested list with predictions from all of the models that were fitted with the \code{multiAPCMC.multifit.object}.
}
\description{
Makes predictions for multiple specifications of an APCMC model at once.
}
\examples{
 multiAPCMC.example.data <- multiAPCMC.example.data
multiAPCMC.example.futuredata <- multiAPCMC.example.futuredata
multifit.example <- multiAPCMC.multifit(multiAPCMC.example.data,
                    startestage=5,
                    vec.link=c('log','power5'),
                    vec.noperiod=c(5),
                    vec.refper=c('outer'),
                    vec.refcoh=c('outer'),
                    nomodelmax=1024)
multiAPCMC.multipred(multifit.example,
                     multiAPCMC.example.futuredata,
                     cuttrend=rep(0,15))

}
