% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.multipredtest.R
\name{multiAPCMC.multipredtest}
\alias{multiAPCMC.multipredtest}
\title{multiAPCMC.multipredtest}
\usage{
multiAPCMC.multipredtest(
  multiAPCMC.multipred.object,
  oos.data,
  validate.what = "oos"
)
}
\arguments{
\item{multiAPCMC.multipred.object}{The list of predictions of the fitted APCMC models that is produced by \code{multiAPCMC.multipred}.}

\item{oos.data}{a dataframe containing information on future periods, cases (e.g. incidence) and person-years (exposed population) in those years. Should have the same resolution as the data on which the models were also fitted (e.g. 1 year age group by 1 year periods, or 5-year age groups by 5-year periods, and so forth) and should have the columns Period, Age, cases, and PY (person-years). In other words, this is the same as the 'futuredata' argument from multiAPCMC.multipred, except with an additional 'cases' column.}

\item{validate.what}{which data should be used to test the predictions against? \code{oos} stands for 'out of sample' and hence only compares fit of predictions to data from the \code{oos.data}. The option \code{not periodfitted} refers to all data that the model was not fitted on. For example, if you have 20 years of data to fit the model on, but had \code{noperiod} of 10, then the first 10 years of that data are used, in addition to the \code{oos.data}, to compare the model against (essentially, you forecast backwards). The option \code{all} compares to everything, including the period that the model was fitted on. Hence, this last option is not recommended.}
}
\value{
returns a nested list with root mean squared error (RMSE) and mean absolute error (MAE) values.
}
\description{
compared model predictions to out of sample data. The function \code{multiAPCMC.multipredtest} is used within the ranktable function, so users are not expected to use this function directly.
}
\examples{
 multiAPCMC.example.data <- multiAPCMC.example.data
multiAPCMC.example.futuredata <- multiAPCMC.example.futuredata
multifit.example <- multiAPCMC.multifit(multiAPCMC.example.data,
                    startestage=5,
                    vec.link=c('log','power5'),
                    vec.noperiod=c(5),
                    vec.refper=c('outer'),
                    vec.refcoh=c('outer'),
                    nomodelmax=1024)
multipred.example <- multiAPCMC.multipred(multifit.example,
                                          multiAPCMC.example.futuredata,
                                          cuttrend=rep(0,15))
multiAPCMC.multipredtest(multipred.example,
                         multiAPCMC.example.futuredata,
                         validate.what="oos")
# note that multiAPCMC.example.futuredata contains 'cases' so that we can use it both to predict
# and to subsequently test the fit of our model to the cases in this future data.
}
