% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.predsummary.R
\name{multiAPCMC.predsummary}
\alias{multiAPCMC.predsummary}
\title{multiAPCMC.predsummary}
\usage{
multiAPCMC.predsummary(
  years,
  modelpred,
  agemin = 0,
  agemax = 120,
  poissondraw = FALSE
)
}
\arguments{
\item{years}{a vector of the years that you want to have summaries for}

\item{modelpred}{the output of a retrievemodel(multipred) or a randpred}

\item{agemin}{the lower bound of the age categories that you want to summarize the outcome over}

\item{agemax}{the upper bound of the age categories that you want to summarize the outcome over}

\item{poissondraw}{set to \code{TRUE} if you want to add a random draw from a Poisson distribution with lambda set to the summary value for that year (e.g. to build a prediction interval). \code{Default is FALSE}.}
}
\value{
returns a nested list with predictions from all of the models that were fitted with the \code{multiAPCMC.multifit.object}.
}
\description{
summarizes the outcome by year within some age range (i.e. adds all incidences together); use this together with randpred to get confidence intervals for an age range (see documentation).
}
\examples{
multiAPCMC.example.data <- multiAPCMC.example.data
multifit.example <- multiAPCMC.multifit(multiAPCMC.example.data,
                                       startestage=5,
                                       vec.link=c('log','power5'),
                                       vec.noperiod=c(5,10),
                                       vec.refper=c('outer','center'),
                                       vec.refcoh=c('outer'),
                                       nomodelmax=1024)
chosenmodel <- multiAPCMC.retrievemodel(multifit.example,
                                       what="fit",
                                       link='power5',
                                       noperiod=10,
                                       refper='center',
                                       refcoh='outer')
years <- 1989:2020
multiAPCMC.predsummary(years,chosenmodel)

}
