% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.reclist.R
\name{multiAPCMC.reclist}
\alias{multiAPCMC.reclist}
\title{multiAPCMC.reclist}
\usage{
multiAPCMC.reclist(dimensions)
}
\arguments{
\item{dimensions}{a vector containing the dimensions of the nested list.}
}
\value{
returns an empty set of nested lists.
}
\description{
makes empty nested lists. This is used by various multiAPCMC functions, but is not expected to be called directly by the user.
}
\examples{

vec.link <- c('power5','poisson')
vec.noperiod <- c(5,10,15,21)
vec.refper <- c('extremes','outer','center', 'first middle','middle last',
                'first second', 'penultimate last')
vec.refcoh <- c('extremes','outer','center', 'first middle','middle last',
                'first second', 'penultimate last')

  depth = c(length(vec.link),
  length(vec.noperiod),
  length(vec.refper),
  length(vec.refcoh))
  multiAPCMC.reclist(depth)

}
