% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.retrievemodel.R
\name{multiAPCMC.retrievemodel}
\alias{multiAPCMC.retrievemodel}
\title{multiAPCMC.retrievemodel}
\usage{
multiAPCMC.retrievemodel(
  multiAPCMC.multipred.object,
  what = "pred",
  link,
  noperiod,
  refper,
  refcoh
)
}
\arguments{
\item{multiAPCMC.multipred.object}{either a multiAPCMC.multipred.object from the multiAPCMC.multipred function, or a multiAPCMC.multifit.object from the multiAPCMC.multifit function.}

\item{what}{should be \code{'pred'} if we want to retrieve predictions from a multiAPCMC.multipred object or it should be \code{'fit'} if we wish to retrieve a model from a multiAPCMC.multifit object.}

\item{link}{should be log-link ("log") or a power5-link ("power5").}

\item{noperiod}{the number of periods in the prediction base.}

\item{refper}{the period references. Options are 'extremes', 'outer', 'center', 'first middle', 'middle last', 'first second', 'penultimate last'.}

\item{refcoh}{the cohort references. Options are 'extremes', 'outer', 'center', 'first middle', 'middle last', 'first second', 'penultimate last'.}
}
\value{
returns a nested list with all of the fitted models.
}
\description{
retrieves a model, or its predictions, with these specifications from either a multiAPCMC.multifit object or from a multiAPCMC.multipred object.
}
\examples{
multiAPCMC.example.data <- multiAPCMC.example.data
multifit.example <- multiAPCMC.multifit(multiAPCMC.example.data,
                                       startestage=5,
                                       vec.link=c('log','power5'),
                                       vec.noperiod=c(5,10),
                                       vec.refper=c('outer','center'),
                                       vec.refcoh=c('outer'),
                                       nomodelmax=1024)
chosenmodel <- multiAPCMC.retrievemodel(multifit.example,
                                       what="fit",
                                       link='power5',
                                       noperiod=10,
                                       refper='center',
                                       refcoh='outer')
                         
                                             
}
