% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.singlefit.R
\name{multiAPCMC.singlefit}
\alias{multiAPCMC.singlefit}
\title{multiAPCMC.singlefit}
\usage{
multiAPCMC.singlefit(multiAPCMC.datashape.object, link = "power5")
}
\arguments{
\item{multiAPCMC.datashape.object}{a data frame containing the variables in the model.}

\item{link}{the specified link function. Choices are log-link ("log") and power5-link ("power5")}
}
\value{
returns a single fitted APCMC model.
}
\description{
Fits multiple a single specification of an APCMC model. This function is called by multiAPCMC.multifit and hence it is not expected that users use the singlefit version directly, even when fitting just one model because multifit allows easy specification of references and such.
}
\examples{
multiAPCMC.example.data <- multiAPCMC.example.data
apcdata <- multiAPCMC.datashape(multiAPCMC.example.data,
                                noperiod=length(unique(multiAPCMC.example.data$Period)),
                                startestage=1,
                                refper1st=1,
                                refper2nd=2,
                                refcoh1st=1,
                                refcoh2nd=2)
multiAPCMC.singlefit(apcdata,link='power5')

}
