% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAPCMC.singlepred.R
\name{multiAPCMC.singlepred}
\alias{multiAPCMC.singlepred}
\title{multiAPCMC.singlepred}
\usage{
multiAPCMC.singlepred(
  multiAPCMC.singlefit.object,
  futuredata,
  noproj = 4,
  cuttrend
)
}
\arguments{
\item{multiAPCMC.singlefit.object}{The list of fitted APCMC models that is produced by \code{multiAPCMC.multifit}.}

\item{futuredata}{a dataframe containing information on future periods and person-years (exposed population) in those years. Should have the same resolution as the data on which the models were also fitted (e.g. 1 year age group by 1 year periods, or 5-year age groups by 5-year periods, and so forth) and should have the columns Period, Age, and PY (person-years).}

\item{noproj}{number of periods to project to the future. Should have at most the length of the number of periods in futuredata.}

\item{cuttrend}{Cut trend in predictions? Should be a vector with at least a length equal to number of periods to predict, e.g. \code{c(0,0,0,0)} or \code{c(0,0.25,0.50,0.75)} or so if we project 4 periods into the future.}
}
\value{
returns a single set of predictions from an APCMC model.
}
\description{
makes a single set of predictions from a fitted multiAPCMC.singlefit object
}
\examples{
multiAPCMC.example.data <- multiAPCMC.example.data
multifit.example <- multiAPCMC.multifit(multiAPCMC.example.data,
                                        startestage=5,
                                        vec.link=c('log','power5'),
                                        vec.noperiod=c(5,10),
                                        vec.refper=c('outer','center'),
                                        vec.refcoh=c('outer'),
                                        nomodelmax=1024)
# if multiAPCMC.ranktable shows that the model with a power5 link function
# a no-period of 10, period references of center and cohort references of 'outer'
# are best, then I could then retrieve this model as follows:
chosenmodel <- multiAPCMC.retrievemodel(multifit.example,
                         what="fit",
                         link='power5',
                         noperiod=10,
                         refper='center',
                         refcoh='outer')
multiAPCMC.singlepred(chosenmodel,
                    multiAPCMC.example.futuredata,
                    noproj=4,
                    cuttrend=rep(0,15))
}
